Remove asusrouter.com in host list
This is a DNS field which is expected to route to the router.
I changed the config so that DNS queries are now passed to the router
(DHCP no longer advertizes the Pi but rather itself). and configured
the Pi as the upstream DNS server for the router. The router
has some defaults related to DNS, on of which is advertising
asusrouter.com to itself. This means it is no longer necessary to
advertise it on the pi itself.