Add restart policy for postgres containers
When restarting the whole Pi I usually run into the exporter failing
to start properly (because Postgres itself isn't running yet) after
which it never retries. This leads to alerts in Grafana going off.

Instead of debugging this issue on every restart this new policy
should take care of it automatically (a sort of self-healing if you
will).