Remove comment on failed backup (Grafana)
The backup did fail in the past but I figured out what the problem was.
The problem was that the cp command in the backup script overwrites
ownership with the user doing the copying (I'm guessing it was root)
whilst the container expected a specific owner (472) within the
container. I ran into the same problem when moving to a custom
mount point on disk.