Remove Postgres user
The user wasn't used just like the Docker group.
Instead root is only allowed to start Docker processes.

It's still a bit weird that the compose files are not in the
root's home but rather in the logged in user's folder.
That is a problem which might be fixed another time.